requirements.txt
fastapi
uvicorn
langchain
chromadb
openai
python-dotenv
langchain-openai
langgraph
grandalf
boto3
pytesseract
pdf2image
pillow
requests